requirements.txt
python-kasa==0.6.2.1